# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vinmo API",
        "description": "Documentação da API do Vinmo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/contracts": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Lista contratos",
                "description": "Display a listing of the resource.",
                "operationId": "971f1fdec41a8eaf0830d94e9ee15496",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Termo de busca",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contratos retornados"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Cria um novo contrato",
                "description": "Store a newly created resource in storage.",
                "operationId": "73a91076ec6423051d986a4149fa8d87",
                "responses": {
                    "200": {
                        "description": "Contrato criado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/contracts/{contract}": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Exibe os detalhes de um contrato",
                "description": "Display the specified resource.",
                "operationId": "3efe2c45b2170cb9f4400d6dd66106fd",
                "parameters": [
                    {
                        "name": "contract",
                        "in": "path",
                        "description": "ID do contrato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do contrato"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Atualiza um contrato",
                "description": "Update the specified resource in storage.",
                "operationId": "3e85ddb8b10c94bb01c376bb06f4be03",
                "parameters": [
                    {
                        "name": "contract",
                        "in": "path",
                        "description": "ID do contrato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contrato atualizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Remove um contrato",
                "description": "Remove the specified resource from storage.",
                "operationId": "d1ff21f53377f46c99d6cf407e68ca14",
                "parameters": [
                    {
                        "name": "contract",
                        "in": "path",
                        "description": "ID do contrato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contrato deletado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/contracts/categories": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Lista as categorias de contrato",
                "operationId": "9ae8f03b8087b524a56f136ad022f86b",
                "responses": {
                    "200": {
                        "description": "Categorias retornadas"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}